refactor: merge mana base recommendations into land base efficiency panel
Move the standalone Mana Base Recommendations panel into the Land Base
Efficiency section as a natural extension of the efficiency scoring.
Removes the separate CollapsiblePanel and nav entry, rendering the
health banner and recommendation rows inline after the factor breakdown.
