+++ src/components/DeckAnalysis.tsx
@@ -22,7 +22,6 @@ import ColorDistributionChart from "@/components/ColorDistributionChart";
 import ManaBaseStats from "@/components/ManaBaseStats";
 import CommanderSection from "@/components/CommanderSection";
 import LandBaseEfficiency from "@/components/LandBaseEfficiency";
-import ManaBaseRecommendations from "@/components/ManaBaseRecommendations";
 import DeckCompositionScorecard from "@/components/DeckCompositionScorecard";
 import HypergeometricCalculator from "@/components/HypergeometricCalculator";
 import PowerLevelEstimator from "@/components/PowerLevelEstimator";
@@ -36,7 +35,6 @@ const ANALYSIS_SECTIONS = [
   { id: "mana-curve", label: "Mana Curve" },
   { id: "color-distribution", label: "Color Dist." },
   { id: "land-efficiency", label: "Land Efficiency" },
-  { id: "mana-recommendations", label: "Mana Recs" },
   { id: "hypergeometric", label: "Draw Odds" },
 ] as const;
 
@@ -234,19 +232,7 @@ export default function DeckAnalysis({
         expanded={expandedSections.has("land-efficiency")}
         onToggle={() => onToggleSection("land-efficiency")}
       >
-        <LandBaseEfficiency result={landEfficiency} />
-      </CollapsiblePanel>
-
-      <CollapsiblePanel
-        id="mana-recommendations"
-        title="Mana Base Recommendations"
-        summary={manaRecommendations.recommendations.length > 0
-          ? `${manaRecommendations.recommendations.length} issue${manaRecommendations.recommendations.length === 1 ? "" : 's'}`
-          : "No issues"}
-        expanded={expandedSections.has("mana-recommendations")}
-        onToggle={() => onToggleSection("mana-recommendations")}
-      >
-        <ManaBaseRecommendations result={manaRecommendations} />
+        <LandBaseEfficiency result={landEfficiency} recommendations={manaRecommendations} />
       </CollapsiblePanel>
 
       <CollapsiblePanel
